feat: update to Vulkan 1.2.189
This is a major refactoring that fixes a number of long-standing issues
with how Vulkan-Docs documentation is parsed.

Instead of using regular expressions for almost everything, the custom
StringConverter and a series of custom inline macros (using
Asciidoctor's Extension API) now perform most of the conversion process
to the LWJGL template format. The formatted content is then patched to
sort the few remaining details (latex math, cleaner struct fields,
unicode unescaping, etc).